commit for version 2.0 finish
as name

+++ Server/DBServer/src/DBApp.cpp
@@ -36,7 +36,7 @@ void CDBServerApp::onExit()
 	LOGFMTI("DBServer ShutDown!");
 }
 
-bool CDBServerApp::onAsyncRequestDelayResp(uint16_t nRequestType, uint32_t nReqSerial, const Json::Value& jsReqContent, uint16_t nSenderPort, uint32_t nSenderID, uint16_t nTargetID)
+bool CDBServerApp::onAsyncRequestDelayResp(uint16_t nRequestType, uint32_t nReqSerial, const Json::Value& jsReqContent, uint16_t nSenderPort, uint32_t nSenderID, uint32_t nTargetID)
 {
 	if ( IServerApp::onAsyncRequestDelayResp(nRequestType, nReqSerial, jsReqContent,nSenderPort,nSenderID,nTargetID) )
 	{
